Doll holding implemented
The player can found a doll in the spots where "thisSpotHasADoll" bool is true.
Then when DollFound() fires, it activates the renderer of the doll in the hand of the player and stores the index of that doll.
When the player hides, the doll renderer is deactivated, and is reactivated once the player stops hidding